+++ src/sis3153_readout_worker.cc
@@ -158,6 +158,8 @@ namespace
                 case CommandType::Accu_Set:
                 case CommandType::Accu_MaskAndRotate:
                 case CommandType::Accu_Test:
+                case CommandType::Accu_Add:
+                case CommandType::Accu_Write:
                     break;
 
                 // Not implemented for the SIS (added in '23 for the MVLC)
@@ -429,6 +431,12 @@ namespace
                 case CommandType::Accu_Set:
                 case CommandType::Accu_MaskAndRotate:
                 case CommandType::Accu_Test:
+                case CommandType::Accu_Add:
+                case CommandType::Accu_Write:
+                case CommandType::MBLTSwappedFifo:
+                case CommandType::Blk2eSST64Fifo:
+                case CommandType::Blk2eSST64SwappedFifo:
+                case CommandType::MvmeRequireVersion:
                     break;
 
                 case  CommandType::Invalid: